+++ Avatar 3.0 VRCSDK/Dependencies/VRChat/Editor/ControlPanel/TGESdkChangelog.cs
@@ -30,7 +30,7 @@ UnityEditor.EditorPrefs.GetFloat("SDKColor_A")
 );
 
         EditorGUILayout.BeginVertical(boxGuiStyle);
-        EditorGUILayout.LabelField("The Black Arms Discord: https://pastebin.com/022rqDb1", EditorStyles.boldLabel);
+        EditorGUILayout.LabelField("The Black Arms Discord: https://discord.gg/m6UfHkY", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("DEDSEC Discord: https://discord.gg/r7RcJCv", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("The Black Arms SDX Dev: PhoenixAceVFX", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("The Black Arms SDX Contributors", EditorStyles.boldLabel);
@@ -57,6 +57,10 @@ My SDX, this decision is final and wont be changed.
 
 This Branch is ENTIRELY SEPERATED from the main branch
 Therefore this will have its own changelog
+
+The Black Arms Avatar 3.0 SDX 1.0.2
+Repaired Discord Links
+Added Guilded.gg Links
 
 The Black Arms Avatar 3.0 SDX 1.0.1
 Updated Discord Links