+++ VRCSDK/Dependencies/VRChat/Editor/ControlPanel/TGELoadBundle.cs
@@ -35,7 +35,7 @@ UnityEditor.EditorPrefs.GetFloat("SDKColor_A")
         EditorGUILayout.BeginHorizontal(boxGuiStyle, GUILayout.Height(26));
         if (GUILayout.Button("The Black Arms Discord"))
         {
-            Application.OpenURL("https://pastebin.com/022rqDb1");
+            Application.OpenURL("https://discord.gg/m6UfHkY");
         }
         if (GUILayout.Button("TGE VRC Asset Server"))
         {
@@ -45,9 +45,9 @@ UnityEditor.EditorPrefs.GetFloat("SDKColor_A")
         {
             Application.OpenURL("https://discord.gg/r7RcJCv");
         }
-        if (GUILayout.Button("VRChat Discord"))
+        if (GUILayout.Button("TBA Guilded"))
         {
-            Application.OpenURL("http://discord.gg/VRChat");
+            Application.OpenURL("https://www.guilded.gg/i/Kk57LVQE");
         }
         EditorGUILayout.EndHorizontal();