+++ VRCSDK/Dependencies/VRChat/Scripts/TGE_Discord.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class TGE_Discord : MonoBehaviour
 {
 
-    private string url = "https://pastebin.com/022rqDb1";
+    private string url = "https://discord.gg/m6UfHkY";
 
     public void OpenURL()
     {