+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 >Any money or profit from my SDX and I will not rebrand  
 >My SDX, this decision is final and wont be changed.  
 
+>The Black Arms SDX 1.10.5
+>Repaired Discord Links
+>Added Guilded.gg Links
+
 >The Black Arms SDX 1.10.4
 >Updated Discord Links for The Black Arms